+++ sismic/model/events.py
@@ -46,11 +46,11 @@ class Event:
 
     def __repr__(self):
         if self.data:
-            return '{}({}, {})'.format(self.__class__.__name__,
-                                       self.name,
-                                       ', '.join('{}={}'.format(k, v) for k, v in self.data.items()))
+            return '{}({!r}, {})'.format(self.__class__.__name__,
+                                         self.name,
+                                         ', '.join('{}={!r}'.format(k, v) for k, v in self.data.items()))
         else:
-            return '{}({})'.format(self.__class__.__name__, self.name)
+            return '{}({!r})'.format(self.__class__.__name__, self.name)
 
 
 class InternalEvent(Event):